Don't use reflection to generate Theseus stuff
